Attention: New file format. CacheWolf can read old format, but old CacheWolf (up to rev. 1786) can't read new format (which is not surprising). (See http://www.geoclub.de/viewtopic.php?f=40&t=33954)

+++ trunk/src/CacheWolf/LogList.java
@@ -92,8 +92,8 @@ public class LogList {
 	 /**
 	  *  Count the number of not-found logs
 	  */
-	 public int countNotFoundLogs() {
-		int countNoFoundLogs = 0;
+	 public byte countNotFoundLogs() {
+		byte countNoFoundLogs = 0;
 		while(countNoFoundLogs < size() && countNoFoundLogs < 5){
 			if(getLog(countNoFoundLogs).getIcon().startsWith("icon_sad")) {
 				countNoFoundLogs++;
